:recycle: simplify save/load_polygon
